Talent-Ops/Talent Operations Chatbot/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
supabase==2.3.4
openai==1.12.0
